requirements.txt
pbr>=0.6,!=0.7,<1.0
anyjson
amqp
kombu
mysql
pillow
python-daemon
requests
flask-restless
oslo.config
oslo.log
